DTO-DAO forgatás

+++ src/main/java/com/springr/first/service/RandomUserServiceImpl.java
@@ -2,14 +2,27 @@ package com.springr.first.service;
 
 
 import com.springr.first.domain.RandomUser;
+import com.springr.first.dto.RandomUser.RandomUserDTO;
+import com.springr.first.exceptions.RandomUserException;
 import com.springr.first.repo.RandomUserRepository;
+import org.modelmapper.ModelMapper;
+import org.modelmapper.TypeToken;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.lang.reflect.Type;
+
 @Service
 @Transactional
-public class RandomUserServiceImpl {
+public class RandomUserServiceImpl implements RandomUserService {
+
+    private final static Type RandomUserListType = new TypeToken<Iterable<RandomUser>>() {
+    }.getType();
+    private final static Type RandomUserDTOListType = new TypeToken<Iterable<RandomUserDTO>>() {
+    }.getType();
+
 
     private RandomUserRepository randomUserRepository;
 
@@ -18,65 +31,99 @@ public class RandomUserServiceImpl {
         this.randomUserRepository = randomUserRepository;
     }
 
+    private ModelMapper modelMapper;
 
-    public RandomUser save(RandomUser randomUser) {
-        return randomUserRepository.save(randomUser);
+    @Autowired
+    public void setModelMapper(@Qualifier("myModelMapper") ModelMapper modelMapper) {
+        this.modelMapper = modelMapper;
     }
 
 
-    public Iterable<RandomUser> save(Iterable<RandomUser> randomUser) {
-        return randomUserRepository.save(randomUser);
+    @Override
+    public RandomUserDTO save(RandomUserDTO randomUser) {
+        return modelMapper.map(randomUserRepository.save(modelMapper.map(randomUser, RandomUser.class)), RandomUserDTO.class);
     }
 
-
-    public RandomUser findOne(Long id) {
-        return randomUserRepository.findOne(id);
+    @Override
+    public Iterable<RandomUserDTO> save(Iterable<RandomUserDTO> randomUserDTOs) {
+        return null;
     }
 
+    @Override
+    public RandomUserDTO findOne(Long id) {
+        return modelMapper.map(randomUserRepository.findOne(id), RandomUserDTO.class);
+    }
 
-    public Boolean exists(Long id) {
+    @Override
+    public boolean exists(Long id) {
         return randomUserRepository.exists(id);
     }
 
+    @Override
+    public Iterable<RandomUserDTO> findAll() {
 
-    public Iterable<RandomUser> findAll() {
-        return randomUserRepository.findAll();
+        System.out.println(randomUserRepository.findAll());
+
+        RandomUser r = randomUserRepository.findAll().iterator().next();
+
+        RandomUserDTO d = modelMapper.map(r, RandomUserDTO.class);
+
+
+        return modelMapper.map(randomUserRepository.findAll(), RandomUserDTOListType);
     }
 
-    public Iterable<RandomUser> findAll(Iterable<Long> id) {
-        return randomUserRepository.findAll(id);
+    @Override
+    public Iterable<RandomUserDTO> findAll(Iterable<Long> ids) {
+        return modelMapper.map(randomUserRepository.findAll(ids), RandomUserDTOListType);
     }
 
-    public Long count() {
+    @Override
+    public long count() {
         return randomUserRepository.count();
     }
 
-
+    @Override
     public void delete(Long id) {
-        randomUserRepository.delete(id);
+        if (exists(id)) {
+            randomUserRepository.delete(id);
+        } else {
+            throw new RandomUserException("Not found");
+        }
     }
 
-
-    public void delete(RandomUser randomUser) {
-        randomUserRepository.delete(randomUser);
+    @Override
+    public void delete(RandomUserDTO randomUserDTO) {
+        randomUserRepository.delete(modelMapper.map(randomUserDTO, RandomUser.class));
     }
 
-
-    public void delete(Iterable<RandomUser> randomUsers) {
-        randomUserRepository.delete(randomUsers);
+    @Override
+    public void delete(Iterable<RandomUserDTO> randomUserDTOs) {
+        randomUserRepository.delete((RandomUser) modelMapper.map(randomUserDTOs, RandomUserListType));
     }
 
-
+    @Override
     public void deleteAll() {
         randomUserRepository.deleteAll();
     }
 
+    // saját logikát ide
+    @Override
+    public Iterable<RandomUserDTO> bulkUpdate(Iterable<RandomUserDTO> randomUserDTOs) {
 
-    public RandomUser update(Long id, RandomUser randomUser){
-        RandomUser old = randomUserRepository.findOne(id);
-        randomUser.setMainKey(id);
-
-        return randomUserRepository.save(randomUser);
+        return null;
     }
+
+
+    @Override
+    public RandomUserDTO updateOne(Long id, RandomUserDTO updateOne) {
+        if (randomUserRepository.exists(id)) {
+            RandomUser old = randomUserRepository.findOne(id);
+            modelMapper.map(updateOne, old);
+            return modelMapper.map(randomUserRepository.save(old), RandomUserDTO.class);
+        } else {
+            throw new RandomUserException("No user found");
+        }
+    }
+
 
 }
